chore(feature-flags): move aggregation resolver out of the JSX constants module
Keep the pure resolveAggregationGroupTypeIndex helper in a plain .ts module so
the non-React kea logic no longer imports through a module that pulls in
@posthog/lemon-ui and JSX. Addresses PR review feedback.

Generated-By: PostHog Code
Task-Id: f32c168e-8eaa-4c82-b53c-0df33261a813
